위치 관련 기능 클래스 생성
* LocationHelper 클래스를 만들어 위치 권한, 위치 정보 취득 부분을 분리함
