+++ app/src/main/java/com/cpstudy/gogumaproject/MainActivity.kt
@@ -57,113 +57,23 @@ class MainActivity : AppCompatActivity(), OnMapReadyCallback {
         binding.lifecycleOwner = this
         setContentView(binding.root)
 
-        LocationServices.getFusedLocationProviderClient(this)
-        checkPermission()
-
         val mapFragment = supportFragmentManager.findFragmentById(R.id.map) as SupportMapFragment
         mapFragment.getMapAsync(this)
 
         binding.btnSuggest.setOnClickListener {
-            myLocation()
+            LocationHelper(this).location {
+                mMap.moveCamera(CameraUpdateFactory.newLatLng(LatLng(it.latitude, it.longitude)))
+                mMap.animateCamera(CameraUpdateFactory.zoomTo(15f))
+            }
         }
     }
 
     override fun onMapReady(googleMap: GoogleMap) {
         mMap = googleMap
 
-        val sydney = LatLng(-34.0, 151.0)
-        mMap.addMarker(
-            MarkerOptions()
-                .position(sydney)
-                .title("Marker in Sydney")
-        )
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(sydney))
-    }
-
-    private fun createLocationRequest() {
-        val locationRequest = LocationRequest.create().apply {
-            interval = 10000
-            fastestInterval = 5000
-            priority = LocationRequest.PRIORITY_HIGH_ACCURACY
-        }
-
-        val builder = locationRequest.let {
-            LocationSettingsRequest.Builder().addLocationRequest(it)
-        }
-
-        val client: SettingsClient = LocationServices.getSettingsClient(this)
-        val task: Task<LocationSettingsResponse> = client.checkLocationSettings(builder?.build())
-        task.addOnSuccessListener {
-            // All location settings are satisfied. The client can initialize
-            // location requests here.
-            fusedLocationClient = LocationServices.getFusedLocationProviderClient(this@MainActivity)
-            checkPermission()
-            fusedLocationClient.requestLocationUpdates(locationRequest, locationCallback, null)
-        }
-
-        task.addOnFailureListener { exception ->
-            if (exception is ResolvableApiException){
-                // Location settings are not satisfied, but this can be fixed
-                // by showing the user a dialog.
-                try {
-                    // Show the dialog by calling startResolutionForResult(),
-                    // and check the result in onActivityResult().
-                    exception.startResolutionForResult(this@MainActivity, 100)
-                } catch (sendEx: IntentSender.SendIntentException) {
-                    // Ignore the error.
-                }
-            }
-        }
-    }
-
-    private val locationCallback: LocationCallback = object : LocationCallback() {
-        override fun onLocationResult(locationResult: LocationResult) {
-            super.onLocationResult(locationResult)
-            val longitude = locationResult.lastLocation.longitude
-            val latitude = locationResult.lastLocation.latitude
-            Log.d("GPS", "$latitude / $longitude")
-            fusedLocationClient.removeLocationUpdates(this)
-            mMap.moveCamera(CameraUpdateFactory.newLatLng(LatLng(latitude, longitude)))
+        LocationHelper(this).location {
+            mMap.moveCamera(CameraUpdateFactory.newLatLng(LatLng(it.latitude, it.longitude)))
             mMap.animateCamera(CameraUpdateFactory.zoomTo(15f))
         }
-
-        override fun onLocationAvailability(locationAvailability: LocationAvailability) {
-            super.onLocationAvailability(locationAvailability)
-
-        }
-    }
-
-    private fun myLocation() {
-        checkPermission()
-        createLocationRequest()
-    }
-
-    private fun checkPermission(): Boolean {
-
-        when {
-            ContextCompat.checkSelfPermission(
-                applicationContext,
-                Manifest.permission.ACCESS_FINE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED -> {
-                Log.d("GPS", "위치 권한 확인")
-                return true
-            }
-            shouldShowRequestPermissionRationale(Manifest.permission.ACCESS_FINE_LOCATION) -> {
-                // In an educational UI, explain to the user why your app requires this
-                // permission for the a specific feature to behave as expected. In this UI,
-                // include a "cancel" or "no thanks" button that allows the user to
-                // continue using your app without granting the permission.
-                Log.e("TAG", "권한을 해야합니다.")
-                Toast.makeText(applicationContext, "권한을 설정해야합니다.", Toast.LENGTH_SHORT).show()
-                requestPermissionLauncher.launch(Manifest.permission.ACCESS_FINE_LOCATION)
-            }
-            else -> {
-                // You can directly ask for the permission.
-                // The registered ActivityResultCallback gets the result of this request.
-                requestPermissionLauncher.launch(Manifest.permission.ACCESS_FINE_LOCATION)
-            }
-        }
-
-        return false
     }
 }
